RDKB-60004:Wi-Fi Telemetry to monitor interop issues with client connectivity

+++ source/dml/tr_181/ml/cosa_wifi_dml.c
@@ -7444,7 +7444,7 @@ AccessPoint_SetParamIntValue
 
     if( AnscEqualString(ParamName, "InteropNumSta", TRUE))
     {
-        if (vapInfo->u.bss_info.inum_sta == iValue)
+        if (vapInfo->u.bss_info.inum_sta == (UINT)iValue)
         {
             return  TRUE;
         }
